feat: implementação da exceção do método delete

+++ java-spring-professional/secao03/desafio/src/main/java/com/gustcustodio/desafio/services/ClientService.java
@@ -11,8 +11,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Optional;
-
 @Service
 public class ClientService {
 
@@ -55,6 +53,10 @@ public class ClientService {
 
     @Transactional
     public void delete(Long id) {
+        if (!clientRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+
         clientRepository.deleteById(id);
     }
